Inject style to hide dialog on unsupported browsers
Unlike Firefox, in WebKit browsers which have open unset
on the dialog won't make it visually hidden.

This is since Firefox has partial dialog support, whereas WebKit has none.

As a result the `close` method could be called when the modal
was meant to be closed / hidden.  This resulted in an error for Safari users.

This is resolved by injecting some CSS
to hide any dialogs that are open.